requirements.txt
pandas
pymongo
pymysql
python-dotenv
sqlalchemy
kagglehub